BVE: use CaDiCaL's elimination score, product plus sum
compute_elim_score() in elim.cpp: elimprod*pos*neg + elimsum*(pos+neg), with
pure literals scoring negative so they are tried first. The sum term breaks
product ties -- 1x10 and 2x5 both give 10, but sums 11 vs 7.

varElimComplexity and the heap comparator become signed to carry the negative
pure-literal scores. Weights are --varelimprod / --varelimsum, both 1.

+++ src/occsimplifier.h
@@ -359,17 +359,17 @@ private:
     /////////////////////
     //Variable elimination
     uint32_t grow = 0; /// maximum grow rate for clauses
-    vector<uint64_t> varElimComplexity;
+    vector<int64_t> varElimComplexity;
     ///Order variables according to their complexity of elimination
     struct VarOrderLt {
-        explicit VarOrderLt(const vector<uint64_t>& _varElimComplexity) :
+        explicit VarOrderLt(const vector<int64_t>& _varElimComplexity) :
             varElimComplexity(_varElimComplexity) {}
 
-        bool operator()(uint64_t x, uint64_t y) const {
+        bool operator()(uint32_t x, uint32_t y) const {
             return varElimComplexity[x] < varElimComplexity[y];
         }
 
-        const vector<uint64_t>& varElimComplexity;
+        const vector<int64_t>& varElimComplexity;
     };
     void        order_vars_for_elim();
     Heap<VarOrderLt> velim_order;
@@ -525,7 +525,7 @@ private:
         , int otherSize
     );
 
-    uint64_t heuristicCalcVarElimScore(const uint32_t var);
+    int64_t  heuristicCalcVarElimScore(const uint32_t var);
     bool resolve_clauses(
         const Watched& ps
         , const Watched& qs
